Remove special handling for "groupBy" field in resolver
The code modification removes the special handling for the "groupBy" field within the `ResolverTypeInterceptor` class. Specifically, it eliminates the block of code that sets a custom resolver (`_groupByFieldResolver`) for the "groupBy" field at the root query level. This means that the "groupBy" field will no longer have a unique resolver assigned to it in this context.