GroupBy type creation and updating connection object with groupby type. (#2541)
## Why make this change?
This pr makes the following changes:
1. Converts the input type objects for aggregations input to enum type
to enable those enums to be used as output as well.
2. Adds in groupby type.

These changes set up the schema appropriately to move to resolver phase
of the execution.

## What is this change?
- `GenerateReturnType` method in `QueryBuilder` updated to include
aggregation fields if `_isAggregationEnabled` is true. This creates the
Connections object and only includes the groupby field if aggregation is
present.
- Added temporary default `_groupByFieldResolver` to
`ResolverTypeInterceptor` to handle `groupBy` operations, which will
throw `NotSupportedException`. The resolver functionality will be added
in following prs.
- `ResolverTypeInterceptor` updated to block `groupBy` operations at the
root query level and set `_groupByFieldResolver` for `GroupBy` type
fields.

## How was this tested?
Unit tests added for all code changes
Query creation experience:
1. Regular select * query:

![image](https://github.com/user-attachments/assets/6c804926-b84c-4411-9285-40ccafdb0ef0)
2. Group by/aggregation experience:

![image](https://github.com/user-attachments/assets/66565407-10ba-4612-92db-351bdcb8a452)
3.Groupby with aggregation added


![image](https://github.com/user-attachments/assets/21378d26-a1f6-4bfa-81a4-4ee139b7ff30)

---------

Co-authored-by: Abhishek  Kumar <102276754+abhishekkumams@users.noreply.github.com>