Refactor HAVING clause logic and update related tests
Refactored the logic for adding the HAVING clause in DWSqlQueryBuilder.cs to ensure it is only added if there are aggregation predicates. Updated related test methods in DwSqlGraphQLQueryTests.cs, GraphQLQueryTestBase.cs, and MsSqlGraphQLQueryTests.cs to reflect changes, switching from sum of prices in stocks_price table to max ID in publishers table.