Adds Query-Identifying IDs to HTTPContext for SQL Queries  (#2553)
## Why make this change?

- We want to be able to associate the statement ID from the SQL query
execution call to the DAB response
- We are now retrieving the statement ID whenever we make a SQL query
execution call and adding it to the httpcontext of our DAB response.
- This change will help DAB users better track DAB performance - by
having access to the statement ID, users can find performance-related
telemetry on the SQL database side, and therefore understand how much
time was spent on the database side vs DAB code.


## What is this change?

- We are creating a new virtual method in QueryExecutor.cs that contains
the logic to create a connection to the data source. This method is
overridden in MsSqlQueryExecutor.cs to include an info message handler
that adds the query identifying IDs to httpcontext.

## Scope of this PR
This change should add the statement ID to the httpcontext for the
following Fabric data sources: SQLDB, Lakehouse and Data Warehouse.

We are still investigating how to enable that for Fabric Datamart, Azure
SQLDB, and MySql and PgSql databases. Those changes - if needed - will
come in a separate PR

## How was this tested?

- [X] Integration Tests
- [X] Unit Tests
- [X] Local testing for Fabric data sources - connected DAB to a Fabric
data source and verified that the query succeeds and that the statement
id was added: works for SQLDB, Lakehouse, Data Warehouse. For Datamart,
the query succeeds as expected, but no statement ID is returned -- we
believe that is a limitation from the Datamart side.
- [X] Local testing - connecting DAB to Azure SQLDB and testing if query
succeeds.

## Sample Request(s)

- Example REST and/or GraphQL request to demonstrate modifications
Screenshot of the httpcontext item added under the header
"QueryIdentifyingIds"

![image](https://github.com/user-attachments/assets/106964b9-98ad-4283-9a5a-f5a70d6edf1b)

---------

Co-authored-by: Giovanna Ribeiro <gribeiro@microsoft.com>