Remove caseInsensitive filter operator from schema generated (#2607)
## Why make this change?
caseInsensitive filter operator is not implemented in the resolvers
leading to runtime error when used. Including it in schema seems to be
an oversight. Removing it for schema generated until we have an approach
defined for case insensitive comparisons with different source types

Relevant issue #2280 

![image5](https://github.com/user-attachments/assets/2876c306-0041-4137-a36a-74b691841434)

## What is this change?
Remove caseInsensitive filter operator from schema generated


## How was this tested?

- [x] Integration Tests
- [x] Unit Tests