Add scalar fields enum generation for GraphQL group-by
Introduce a new feature to generate a scalar fields enum for group-by arguments in GraphQL queries by adding a call to `EnumTypeBuilder.GenerateScalarFieldsEnumForObjectType` in `GraphQLSchemaCreator.cs`. If the enum is successfully created, it is added to the `objectTypes` collection.

Update `TryGenerateEnum` in `EnumTypeBuilder.cs` to return a boolean indicating whether the enum type was generated.

Rename and make `ValidateAggregationNodeIsNotGeneratedWhenNoNumericFieldsPresent` method virtual in `GraphQLQueryTestBase.cs` to `TestNoAggregationOptionsForTableWithoutNumericFields`, allowing derived test classes to override it.

Update derived test classes (`DwSqlGraphQLQueryTests.cs`, `MsSqlGraphQLQueryTests.cs`, `MySqlGraphQLQueryTests.cs`, and `PostgreSqlGraphQLQueryTests.cs`) to override the `TestNoAggregationOptionsForTableWithoutNumericFields` method and call the base implementation.