Remove "And", "Or", and Relationships as valid OrderByInput types. (#2625)
## Why make this change?

Closes https://github.com/Azure/data-api-builder/issues/2286

## What is this change?

The `InputTypeBuilder` class holds a function called
`GenerateOrderByInputTypeForObjectType`, which is responsible for
populating the GraphQL Schema with the valid OrderByInput types. It was
mistakenly adding in "and", "or", and columns that are in entities
associated through relationships as valid OrderByInput types. "and" and
"or" will never be valid input types, and we do not yet support order by
using columns from related entities.

## How was this tested?

Manually verified that the correct error was returned when trying to use
the incorrect orderBy queries, and manually verified that we no longer
have them in the schema.

Added regression test to verify we get an error when attempting these
queries.

## Sample Request(s)


These are some example graph QL queries that are invalid but were
showing as valid in our schema without this fix.



>publishers (first: 5 orderBy:  {books:  {  title: DESC }}){
>  items {
>    id
>     books {
>       items {
>        title
>      }
>   }
> }
> }


>books(orderBy: { or: { id: ASC, title: ASC } }) {
    items {
      id
    }
  }

---------

Co-authored-by: Aniruddh Munde <anmunde@microsoft.com>