E2E Health Endpoint Execution changes (#2614)
## Why make this change?
Closes  #2613 #2557 #2532 

## What is this change?
This PR includes changes in `ComprehensiveHealthReportResponseWriter` to
call the `HealthCheckHelper` and pass the RuntimeConfig. This
`HealthCheckHelper` internally executes the queries and forms the
`ComprehensiveHealthCheckReport`.

In the `HealthCheckHelper` file, we have the following changes
1. Calls a function to fetch the base URL of the hosted env from the
HttpContext.
2. Calls a function to update the basic details like version, app-name
and configuration
3. Further it updates the health check details  [Explained Below]
4. Finally based on the status of all the health checks carried out -
Updates the global health status as Healthy in case all are Healthy and
Unhealthy in case any single is Unhealthy.

Health Check
This section has two checks - one for data source and the other for all
entities.
Important point to note - In case Health is null in the config - health
is considered enabled.

1. Data source: In this, we have basic defined queries (given in design
doc) which are run based on the type of database. We compare the
response time on this query with the threhold given to check health
status.
2. Entities: In this we traverse over all entities, in case health is
globally enabled in runtime config.
We then check for each entity, if rest is globally enabled and rest is
enabled at entity level. If so, we run the rest check for this entity.
Similarly we check if graphql is globally enabled and graphql is enabled
at entity level, we run the graphql check for this entity. We traverse
over this entire entities array to get the health status for rest and
graphql endpoints.
Note: Only if enabled: false in th config is the entity excluded from
the comprehensive health report.

REST CHECK
In case of rest check, we form the URL which is
<baseURL><restpath>/<EntityName>$first=<firstValue>
Now we execute a GET query on the rest endpoint of our engine to check
the response time of this query to compare with threshold.

GRAPHQL CHECK
In case of graphql check, we first fetch the metadata provider for the
particular database type and using this metadata provider we get the
column names in all the entities.
After fetching the column names we create the graphql query like this:
{"query":"{publishers(first:4) {items {id name} }}"}
The we excute a POST request on the graphql endpoint of our engine to
get the response time to compare with threshold.

## How was this tested?

- [ ] Integration Tests
- [x] Unit Tests
- [x] Manual Tests

## Sample Request(s)

Sample Report is attached in the PR
<img width="281" alt="image"
src="https://github.com/user-attachments/assets/e8dd922f-9742-4646-b0e0-3143c2a74ec0"
/>

[ComprehensiveHealthReportResponse.json](https://github.com/user-attachments/files/19183696/ComprehensiveHealthReportResponse.json)

---------

Co-authored-by: sezalchug <sezalchug@microsoft.com>
Co-authored-by: Aniruddh Munde <anmunde@microsoft.com>