Enable Custom log-level for Classes via log-level Filters (#2620)
## Why make this change?

This change solves tasks, #2584, #2585, #2586, #2587, #2588, #2569. 
All of them are part of the bigger task of #2563 which adds filters for
logging.

## What is this change?
This change allows the users to add more specific logging filters to
their config file, this means that log-levels meant for a specific class
will take priority over the default log-level.
In the following example, even if the default is in `debug`, the logging
for `IQueryExecutor` will be in `information`.
```
"Azure.DataApiBuilder.Core.Resolvers.IQueryExecutor": "information",
"default": "debug"
```
Specific Changes:
- Changed the schema file so that it allows for different keywords that
represent the classes, allowing each of them to have a different
log-level. It also allows the user to add a general log-level through
the use of the keyword `default` 
- Updated DAB Validate so that it checks if the keywords used in the
config file are valid classes that we use to give logging information to
the users.
- Change logic that deserializes the new properties in the config file
and stores them inside of the `TelemetryOptions` class and it uses the
saved log-levels to dictate if a specific logger uses a different
log-level from the default. This change gets rid of the `loggerFactory`
and changes it to a dictionary.

## How was this tested?
- [ ] Integration Tests
- [X] Unit Tests

Following tests were added:
- Test that filters are deserialized correctly and their values are what
we expect them to be.
- Test that invalid filters throw an error.
- Test that between two filters, the one that is more specific takes
priority.

## Sample Request(s)

![image](https://github.com/user-attachments/assets/4dbf7a98-7aba-4781-97f3-4b32fa8d77e0)

---------

Co-authored-by: Ruben Cerna <rcernaserna@microsoft.com>