Hot Reload for Loggers (#2634)
## Why make this change?
Closes issues #2580 and #2582

## What is this change?
Adds hot reload feature for loggers which makes it possible for the user
to change the log level filters without restarting the application. This
is done by ensuring that the logger factories are updated when there is
a hot reload event.
This feature is also available in production mode, it is the only hot
reloadable property in production mode.

## How was this tested?
- [ ] Integration Tests
- [x] Unit Tests
- [X] Manual Testing

Test added:
- Check if log-level value is properly updated with hot reload, by
checking if it is different from previous value and if it is the
expected value.

## Sample Request(s)
Before hot reload:

![image](https://github.com/user-attachments/assets/70f3c3e7-936c-44df-85ba-a6c1f9b08dd7)

![image](https://github.com/user-attachments/assets/daffabb4-f542-4bdc-8d1b-56a3aa7961c8)

After hot reload:

![image](https://github.com/user-attachments/assets/8bc14fce-ef8a-4b4e-99b0-16f81bb0e04e)

![image](https://github.com/user-attachments/assets/f6ba26ea-720c-40de-983b-f33db969d349)

---------

Co-authored-by: Ruben Cerna <rcernaserna@microsoft.com>