Fixing Pipelines due to Ubuntu 24.04 update removing Mono (#2648)
## Why make this change?
Resolves #2649

## What is this change?
The pipelines are failing due to udpate from ubutu 24.04.
We are using the latest version in our yamls hence, this change needs to
be handled.
<img width="527" alt="image"
src="https://github.com/user-attachments/assets/b1f14841-ea20-4b18-8b17-b9f1c43b4022"
/>

The error message we are seeing is due to the fact that default Ubuntu
24.04+ image on Microsoft-hosted agents does not include Mono, which is
required for the NuGet client that powers NuGetCommand@2 in our
pipeline. Microsoft recommends migrating to using the
NuGetAuthenticate@1 task and .NET CLI commands instead of the older
NuGetCommand@2 task, which relies on Mono.

Steps to Fix the Issue:
- Remove NuGetCommand@2 task: Since this task requires Mono (which is no
longer available by default on Ubuntu 24.04+), we should remove it and
replace it with .NET CLI commands for restoring the NuGet packages.
- Add .NET CLI restore command: Instead of using the NuGetCommand@2
task, use the DotNetCoreCLI@2 task to restore NuGet packages.
- Update NuGet authentication: Ensure that you're using
NuGetAuthenticate@1 before the restore task to handle authentication.

## How was this tested?
The pipelines are running as expected.

---------

Co-authored-by: sezalchug <sezalchug@microsoft.com>