Include cache options in the add and update verbs of CLI (#2655)
## Why make this change?

closes https://github.com/Azure/data-api-builder/issues/2603

## What is this change?

Includes caching options with the `add` and `update` verbs for the CLI.
These verbs are used to add or update new entities in a config file, and
will now include support for the `cache` entry in those entities.

## How was this tested?

Current test suite.

## Sample Request(s)

You can test with the CLI by using the `add` or `update` verb along with
the flags `--cache.enabled` and `--cache.ttl`, default values are false
and 5.

---------

Co-authored-by: Ruben Cerna <rcernaserna@microsoft.com>
Co-authored-by: Aniruddh Munde <anmunde@microsoft.com>