Migrate from JsonSchema.NET to Newtonsoft.Json.Schema
JsonShema.NET supports the `oneof + const` validation which NJsonSchema didn't, which is great. Sadly JsonSchema.NET returns tons of errors about things that aren't really errors. An example is the if statement for cosmosdb that requires the database/container/schema properties. If the user has selected mssql, they haven't selected cosmos, and so all the IF statements for cosmosdb are FALSE.

This makes error handling quite difficult.

My workaround is to switch to another popular library: Newtonsoft.Json.Schema. We're already using Newtonsoft.Json in some parts, so this seems to be a better fit.

This also returns the LineNumber validation, which is great as this requires less changes in the PR overall!