Add Level 2 Caching (#2619)
## Why make this change?

Work for #2543 (Add Level 2 Caching).

## What is this change?

First code pushed on this branch.

Summary of the changes:
- updated package reference for FusionCache from `1.0.0` to `2.1.0`
- added package references to
`ZiggyCreatures.FusionCache.Serialization.SystemTextJson`,
`Microsoft.Extensions.Caching.StackExchangeRedis` and
`ZiggyCreatures.FusionCache.Backplane.StackExchangeRedis` to support
L2+backplane on Redis
- added support for L2+backplane in the initial config (`Program.cs`)
- added `RuntimeCacheOptions` on top of `EntityCacheOptions`:
`EntityCacheOptions` was used to, I suppose, avoid having 2 similar
classes since they had the same shape, but now that is not true anymore
- added initial version of `RuntimeCacheLevel2Options` to contain the
related config values
- aligned some tests that broke because they were directly instantiating
and passing `EntityCacheOptions` instead of the new
`RuntimeCacheOptions`
- added new support classes (eg: `RuntimeCacheOptionsConverterFactory`)
- added `CACHE_ENTRY_TOO_LARGE` message instead of the previously used
`CACHE_KEY_TOO_LARGE` one, which was used when the entry was too large
(not the key)
- changed the generic `T` in `DabCacheService.GetOrSetAsync`, from
`JsonElement` (same name but a totally different thing than
`JsonElement` class) to `TResult` (which was already used in the other
`GetOrSetAsync` overload, to avoid confusion
- fixed some typos (eg: "ommitted" instead of "omitted" etc)
- added some temporary `// TODO` comments to get back to them later

Still missing:
- although I already worked on the config generation side of things (eg:
not just reading the config, but also generating it via the CLI), I
think I haven't finished it yet
- there are also some points where I'm not sure the convention I
extrapolated is fully correct (eg: a config sub-object should be
generated only if it's not `null` or should we do a comparison with the
inner default values to see if we should skip it as a whole?)
- since DAB is using FusionCache, it's also using
[Auto-Recovery](https://github.com/ZiggyCreatures/FusionCache/blob/main/docs/AutoRecovery.md):
becaus of this, I'd like to tweak some of the default settings (eg:
allow background distributed operations to make things faster, etc). But
we'd have to first check if we want to expose some options to drive this
or we go full-auto-mode

Not (necessarily) part of this effort, but already mentioned and to keep
in mind:
- avoid calculating each cache entry size, since SizeLimit is not
currently used, nor it is possible to use it via some config (so, for
now, it's useless)
- start thinking about enabling some resiliency features like
[Fail-Safe](https://github.com/ZiggyCreatures/FusionCache/blob/main/docs/FailSafe.md),
[Soft
Timeouts](https://github.com/ZiggyCreatures/FusionCache/blob/main/docs/Timeouts.md),
etc
- in the future it may be nice to introduce
[Tagging](https://github.com/ZiggyCreatures/FusionCache/blob/main/docs/Tagging.md)
to automatically evict all the cached entries containing data about an
entity after an update to said entity. It would be great

## How was this tested?

Tests run locally:
- [x] Unit Tests

## Sample Request(s)

N/A

---------

Co-authored-by: aaronburtle <93220300+aaronburtle@users.noreply.github.com>
Co-authored-by: aaron burtle <aaronburtle@microsoft.com>
Co-authored-by: Aniruddh Munde <anmunde@microsoft.com>
Co-authored-by: RubenCerna2079 <32799214+RubenCerna2079@users.noreply.github.com>