Write CacheTtl to entity when add/update command is used and CacheTtl is provided (#2717)
## Why make this change?

Closes https://github.com/Azure/data-api-builder/issues/2716

## What is this change?

When we create the entity cache options, we start with a new object and
then later set the values for the fields. Because of this, the `bool`
`UserProvidedTtlOptions` must explicitly be set when we later set these
values or it will forever remain `false`. When it is false it is not
written back out into the configuration file.

## How was this tested?

Manually used the update command in the CLI to verify the field is
written out.

## Sample Request(s)

- Example REST and/or GraphQL request to demonstrate modifications
- Example of CLI usage to demonstrate modifications