Cherry-Pick Bug Fix for 1.5 Release (#2718)
## Why make this change?

When we create the entity cache options, we start with a new object and
then later set the values for the fields. Because of this, the `bool`
`UserProvidedTtlOptions` must explicitly be set when we later set these
values or it will forever remain `false`. When it is false it is not
written back out into the configuration file.

## What is this change?

Fixes bug found in release candidate for 1.5
Cherry-picked PR:
- #2717 

## How was this tested?

## Sample Request(s)

Co-authored-by: aaronburtle <93220300+aaronburtle@users.noreply.github.com>