Add authentication validation to the JSON schema (#2646)
## Why make this change?

Closes #2643 

The JSON schema does not contain a list of auth providers, nor does it
validate the use of the `jwt` property. This results in a suboptimal
developer experience. By adding this list and validation, developers
have an easier time learning about all the options and can be more
productive.

## What is this change?

- This change modifies the JSON schema with a list of providers and
validation of the providers and the `jwt` property. [StackOverflow
helped me](https://stackoverflow.com/q/79557612/3013479) with the finer
details of the validation.
  - I also added tests to validate my changes.
- Migrate from `NJsonSchema` to `Newtonsoft.Json.Schema` because
`NJsonSchema` [doesn't validate `anyof + const`
correctly](https://github.com/RicoSuter/NJsonSchema/issues/1596).
`Newtonsoft.Json.Schema` does work in this case.
- Changed `https` to `http` in the schema file to be spec compliant.
Draft-07's original URL is http-based.
- If you try to validate a document with the `https` url, it will fail
because of the unknown schema URL.
- This does **not** impact security. The schema should just be a
placeholder for the json validator binary to know what rules to use, and
it's redirected to https when you navigate to it, anyway.

## How was this tested?

- [x] Integration Tests
- [x] Unit Tests

## Sample Request(s)


![image](https://github.com/user-attachments/assets/3d70b56f-78e4-4f10-90ab-4f8a1356ce6c)


![image](https://github.com/user-attachments/assets/37b11c95-d443-4ee2-81c2-a54365a47cbe)


![image](https://github.com/user-attachments/assets/a04030cf-cec8-41cc-80a3-78e17f155bb6)

---------

Co-authored-by: Aniruddh Munde <anmunde@microsoft.com>
Co-authored-by: RubenCerna2079 <32799214+RubenCerna2079@users.noreply.github.com>