Allow unsupported cache-control options without error (#2737)
## Why make this change?

Closes https://github.com/Azure/data-api-builder/issues/2735

## What is this change?

Removes the check to make sure that only supported `cache-control`
options are included in the request headers.

## How was this tested?

Manually ran GraphQL request with headers that failed in bug report. To
reproduce use any valid graphQL query with cache enabled in the runtime
section of the config, and for the entity being queried.


![image](https://github.com/user-attachments/assets/942e6870-e5a9-4bcd-86a0-200a3ad7a0b9)

Then any graphQL query against that entity with `cache-control` headers
should enter the code path in question.



## Sample Request(s)

```
query {
  publishers (first: 5 orderBy: {name: DESC
  })
  {
    items {
      id
    }
  }
}
```