Change property name of log-type to dab-identifier for Azure Log Analytics (#2805)
## Why make this change?
- This change closes issue #2804 
The property `log-type` had a change in function, which also means that
there needs to be a change in name.

## What is this change?
Only changes the name of the property from `log-type` to
`dab-identifier` since the changes in logic related to the changes in
function were already implemented beforehand.

## How was this tested?

- [ ] Integration Tests
- [ ] Unit Tests
- [X] Manual Tests
Tested if the new property is serialized and deserialized correctly and
if `configure` command works as expected with new property, while
leaving no traces of the old property.

## Sample Request(s)
--runtime.telemetry.azure-log-analytics.dab-identifier