Support for relative nextLink using `next-link-relative` property for Reverse Proxy Scenarios (#2788)
## Why make this change?

- Closes #2677  
When Data API builder (DAB) is deployed behind a reverse proxy that
handles SSL termination (converting https to http), the pagination
nextLink URL in REST responses was generated with the http scheme
instead of https. This resulted in broken next page links for clients
accessing the service securely. The reported bug highlighted that the
nextLink should use the original client-facing scheme and host, or
provide a relative/schema-less link, or leverage the x-forwarded-proto
and x-forwarded-host headers set by the reverse proxy.

## What is this change?

- Users can also optionally set `x-forwarded-proto` and
`x-forwarded-host headers` as needed to match their deployment
environment. If not set, it will pick the default path from the request
context.
- Also refer
[X-Forwarded-Proto](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-Forwarded-Proto)
and
[X-Forwarded-Host](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/X-Forwarded-Host).

## How was this tested?

- [ ] Integration Tests
- [x] Unit Tests

## Sample Request(s)

- GET `http://localhost:5000/api/<your_entity>`