Add logic to send logs through File Sink (#2825)
## Why make this change?

- Closes issue #2578
- In order to complete the File Sink feature, the logic that sends the
logs created by DAB to the file is required.

## What is this change?

- This change implements the logic that sends all the logs from DAB to
the file through the path the user requests. This is done with the help
of Serilog which is a logging library that simplifies the creation of
file sinks.
- The `Startup.cs` program creates the Serilog logger pipeline and adds
it as part of the services so that it is used later by the `Program.cs`
to set the different loggers with the Serilog pipeline and allow the
logs to be sent to the file sink.
- We also deleted the `RollingIntervalMode.cs` since we discovered that
Serilog has its own rolling interval enum class, which makes the one
implemented in DAB obsolete.

## How was this tested?

- [ ] Integration Tests
- [X] Unit Tests

Created tests that check if the services needed for the File Sink exist
when the File Sink property is enabled.
Also, created test to check if the file sink with the appropriate name
is created when the property is enabled.

## Sample Request(s)
<img width="917" height="136" alt="image"
src="https://github.com/user-attachments/assets/a470844a-6ea3-4d05-9dfa-650f878c1c59"
/>
<img width="1873" height="948" alt="image"
src="https://github.com/user-attachments/assets/1895eaae-0223-4d48-924e-c5275a98e3dc"
/>