Ensure JsonDocuments cleaned up after request (#2865)
## Why make this change?
Closes #2858

When stored procedures are applied with authorization policies in DAB
configuration, then at runtime we observe that queries or mutations via
stored procedures fail unexpectedly with object disposed exception. This
PR fixes the bug.

In the query and mutation root resolvers, the intermediate results in
the form of JsonDocuments are registered for cleanup. These were getting
cleaned even before the leaf resolvers executed so DAB would encounter
exception accessing already disposed objects.

If registerForCleanup is invoked without additional parameter specifying
when to cleanup, then by default, the cleanup happens immediately after
the current resolver execution. This was causing the objects to be
disposed early for stored procedure flow before the leaf resolvers were
executed.

## What is this change?

Invoke registerForCleanup with the additional param specifying the
cleanup to happen after request is complete.

## How was this tested?

- [x] Integration Tests
- [ ] Unit Tests