Add MCP Runtime serialization/deserialization and fix tests related to it (#2881)
## Still missing
This PR is still missing some test changes as well as some changes to
the `McpRuntimeOptions` file to properly serialize/deserialize the
properties found inside `mcp`.

## Why make this change?

- Tests related to the serialization and deserialization of the MCP
properties were failing or missing.

## What is this change?

- DAB is now able to serialize and deserialize the MCP properties, this
is done through the addition of the following files:
- `McpRuntimeOptionsConverterFactory` reads and writes the properties
that are found inside of the `mcp` property.
- `McpDmlToolsOptionsConverter` reads and writes the properties that are
found inside the `dml-tools` property.
- Fixed the `McpRuntimeOptions`, `RuntimeConfig`, and `RuntimeOptions`
files to allow for proper serialization and deserialization and reflect
the changes related to the new `mcp` property.
- Separated `McpDmlToolsOptions` class from the `McpRuntimeOptions` file
where it originally resided.

- Fixed tests that were failing or missing changes for the new `mcp`
property:
- `ModuleInitializer` for both CLI and Service tests now ignore
`IsMcpEnabled` property which is only used inside of DAB and shouldn't
be written inside the config file.
- `RuntimeConfigLoaderJsonDeserializerTests` changed tests to include
`mcp` property.

## How was this tested?

- [ ] Integration Tests
- [X] Unit Tests